restructed models into own lib
